Move some logic using rustc datastructures to the `rustc_smir` module
